Added source files and updated common project files.
The source files contain almost the entirety of the Coloring Book app.
There are a few things that would need to be implemeneted (or
alternatively, removed) for a clean build. So as of this commit,
the application isn't automatically buildable when the repository is
cloned. This will need to be addressed.

The common project files have been updated to reflect additional
license requirements.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>