Added the initial project build.sh script.
The build script prepares a freshly cloned repository for compilation.
There are several custom components the application uses that cannot be
version controlled, because they are specific to the developer(s). This
includes things like public keys tied to a Google account. Thus, the
build script assists developers in getting this sorted out so they can
successfully compile the application.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>