Optimized the flood-fill algorithm.
Instead of using ArrayLists, which can dynamically allocate memory up to
arbitrary size and cause OOM errors, the algorithm is now using boolean
arrays that are preallocated with the maximum width and height of the
images dimensions. The boolean arrays are multidimensional and default
to false values. When a pixel is to be filled with a color, the
corresponding boolean[x][y] value is set to true.

This is faster than using ArrayLists and ultimately uses much less
heap memory.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>