Removed unnecessary import.
This import was never used. It was intended to be used by a dynamic
database-path getter. This is because different devices may have
different database paths. However, a different approach to database
installation has been taken. Instead of copying the .db file to this
"dynamic" path (the path can vary by device), the database itself and
all database updateds occur through execution of DDL statements via
changescripts.

Thus, this import is not necessary now, nor has it ever been in the
past. It just happened to get committed.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>